requirements.txt
opencv-python
pillow
ultralytics
djitellopy
python-telegram-bot
tkintermapview
tk
asyncio
